requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
